+++ lib/pry/commands/disabled_commands.rb
@@ -1,2 +1,0 @@
-Pry::Commands.disabled_command("edit-method", "Use `edit` instead.")
-Pry::Commands.disabled_command("show-command", "Use show-source [command_name] instead.")